Swapped old trees for new ones, and added forest.
Added forest with newly updated trees. Also added some rocks to make it possible to climg up on the rocks where the tower is located in field area. Slightly scaled Flacan's rock spiral to check if it made it easier to climb - ended up not checking if it made it easier, but now have it added to list of things to be done, just like the bug where you can get trapped in between the inside of the cave and the outside of the mountain on the path down from hightown to the forest.